Move manifest create w/ external items to own file
